replacing np.int as well

+++ tvb_contrib/tvb/contrib/scripts/utils/data_structures_utils.py
@@ -61,13 +61,13 @@ class CalculusConfig(object):
 
 def is_numeric(value):
     return isinstance(value, (float, np.float, np.float64, np.float32, np.float16, np.float128,
-                              int, np.int, np.int0, np.int8, np.int16, np.numpy.int32, np.int64,
+                              int, int, np.int0, np.int8, np.int16, np.numpy.int32, np.int64,
                               complex, np.complex, np.complex64, np.complex128, np.complex256,
                               np.long, np.number))
 
 
 def is_integer(value):
-    return isinstance(value, (int, np.int, np.int0, np.int8, np.int16, np.numpy.int32, np.int64))
+    return isinstance(value, (int, int, np.int0, np.int8, np.int16, np.numpy.int32, np.int64))
 
 
 def is_float(value):
@@ -626,7 +626,7 @@ def assert_equal_objects(obj1, obj2, attributes_dict=None, logger=None):
 def shape_to_size(shape):
     shape = np.array(shape)
     shape = shape[shape > 0]
-    return np.int(np.max([shape.prod(), 1]))
+    return int(np.max([shape.prod(), 1]))
 
 
 def shape_to_ndim(shape, squeeze=False):
@@ -662,7 +662,7 @@ def assert_arrays(params, shape=None, transpose=False):
     # type: (object, object) -> object
     if shape is None or \
             not (isinstance(shape, tuple)
-                 and len(shape) in range(3) and np.all([isinstance(s, (int, np.int)) for s in shape])):
+                 and len(shape) in range(3) and np.all([isinstance(s, (int, int)) for s in shape])):
         shape = None
         shapes = []  # list of all unique shapes
         n_shapes = []  # list of all unique shapes' frequencies
@@ -769,14 +769,14 @@ def make_int(x, precision="64"):
         elif isequal_string(precision, "32"):
             return x.astype(np.numpy.int32)
         else:
-            return x.astype(np.int)
+            return x.astype(int)
     else:
         if isequal_string(precision, "64"):
             return np.int64(x)
         elif isequal_string(precision, "32"):
             np.numpy.int32(x)
         else:
-            return np.int(x)
+            return int(x)
 
 
 def copy_object_attributes(obj1, obj2, attr1, attr2=None, deep_copy=False, check_none=False):
